moved ComponentManifest class from 'internal.core.deployment.parser' to 'api.config' package
