+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/core/deployment/AppCreator.java
@@ -23,6 +23,7 @@ import com.google.common.collect.HashMultimap;
 import com.google.common.collect.ListMultimap;
 import com.google.common.collect.SetMultimap;
 import org.wso2.carbon.uuf.api.Placeholder;
+import org.wso2.carbon.uuf.api.config.ComponentManifest;
 import org.wso2.carbon.uuf.api.config.Configuration;
 import org.wso2.carbon.uuf.api.reference.AppReference;
 import org.wso2.carbon.uuf.api.reference.ComponentReference;
@@ -43,7 +44,6 @@ import org.wso2.carbon.uuf.exception.InvalidTypeException;
 import org.wso2.carbon.uuf.exception.MalformedConfigurationException;
 import org.wso2.carbon.uuf.exception.UUFException;
 import org.wso2.carbon.uuf.internal.core.auth.SessionRegistry;
-import org.wso2.carbon.uuf.internal.core.deployment.parser.ComponentManifest;
 import org.wso2.carbon.uuf.internal.core.deployment.parser.ComponentManifestParser;
 import org.wso2.carbon.uuf.internal.core.deployment.parser.ConfigurationParser;
 import org.wso2.carbon.uuf.internal.core.deployment.parser.DependencyNode;
